feat(chat): 멘티관리에 채팅 종료 버튼(공유 ChatRoomView 추출) + 멘티관리 플로팅 숨김
- ChatRoomView: 종료 바+스레드+입력+확인 다이얼로그를 공유 컴포넌트로 추출
  → ChatModal과 멘토 멘티관리 페이지가 동일 동작/디자인 사용(종료 버튼·상대 종료 강조)
- 멘티관리: 좌측 멘티 목록 유지 + 우측에 ChatRoomView(종료 가능) 임베드
- MentorChatLauncher: /feedback/live-mentee 라우트에선 플로팅 미노출(페이지 자체 목록 존재)
